# Speed up pack action registration through the /v1/packs/register API endpoint