--- a/st2common/st2common/bootstrap/base.py
+++ b/st2common/st2common/bootstrap/base.py
@@ -54,21 +54,29 @@ EXCLUDE_FILE_PATTERNS = [
 class ResourceRegistrar(object):
     ALLOWED_EXTENSIONS = []
 
-    def __init__(self, use_pack_cache=True, fail_on_failure=False):
+    def __init__(self, use_pack_cache=True, use_runners_cache=False, fail_on_failure=False):
         """
         :param use_pack_cache: True to cache which packs have been registered in memory and making
                                 sure packs are only registered once.
         :type use_pack_cache: ``bool``
 
+        :param use_runners_cache: True to cache RunnerTypeDB objects in memory to reduce load on
+                                  the database.
+        :type use_runners_cache: ``bool``
+
         :param fail_on_failure: Throw an exception if resource registration fails.
         :type fail_on_failure: ``bool``
         """
         self._use_pack_cache = use_pack_cache
+        self._use_runners_cache = use_runners_cache
         self._fail_on_failure = fail_on_failure
 
         self._meta_loader = MetaLoader()
         self._pack_loader = ContentPackLoader()
         self._runner_loader = RunnersLoader()
+
+        # Maps runner name -> RunnerTypeDB
+        self._runner_type_db_cache = {}
 
     def get_resources_from_pack(self, resources_dir):
         resources = []