This pull request speeds up registering pack actions through the ``/v1/packs/register`` API endpoint (e.g. when using `st2 pack install` command).

We do that by caching `RunnerTypeDB` objects in memory for each "register session" (aka each time user hits ``/v1/packs/register`` API endpoint). Since we re-initialize the cache on each API request, this means there are no issues with potential stale cache, but we still see benefits when registering packs with a lot of resources.

This helps in situations where pack has a lot of actions such as the ``aws`` pack. Previously RunnerTypeDB was retrieved from the database ``N`` times where ``N`` is number of actions in a pack and now it's only retrieved ``M`` times where ``M`` is number of unique runners referenced by actions in a particular pack (in ``aws`` pack that's 1 - Python runner).

Some hard data.

Before this change:

```bash
2018-09-14 10:29:23,180 INFO [-] 8f73eafd-f38c-432e-9156-ca7f1d8f67ae - 200 194 114645.359ms (content_length=194,request_id='8f73eafd-f38c-432e-9156-ca7f1d8f67ae',runtime=114645.359,remote_addr='127.0.0.1',status=200,method='POST',path='/packs/register')
```

After this change:

```bash
2018-09-14 10:31:51,825 INFO [-] 2860c91e-f93e-48df-ba1a-465d5bf3f4ed - 200 194 96030.742ms (content_length=194,request_id='2860c91e-f93e-48df-ba1a-465d5bf3f4ed',runtime=96030.742,remote_addr='127.0.0.1',status=200,method='POST',path='/packs/register')
```
So around ~18 seconds improvement.

96 seconds is still a very long time, but the whole platform was not designed around "packs with many actions" in mind. Just reading and validating all the action metadata from disk takes around ~60 seconds (not taking into account saving this metadata to the database).